app/requirements.txt
streamlit==1.42.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.1
